+++ lib/ic/endpoints_test.go
@@ -72,11 +72,6 @@ var (
 		"/scim/v2/{realm}/Me",
 		"/scim/v2/{realm}/Users",
 		"/scim/v2/{realm}/Users/{name}",
-		"/identity/scim/v2/{realm}/Groups",
-		"/identity/scim/v2/{realm}/Groups/{name}",
-		"/identity/scim/v2/{realm}/Me",
-		"/identity/scim/v2/{realm}/Users",
-		"/identity/scim/v2/{realm}/Users/{name}",
 
 		// administration endpoints
 		"/identity/v1alpha/{realm}",